ci: change integration tests to options-only
